Move KtorDroidKaigiApi to common class of MultiPlatformProject
