+++ data/api-impl/src/commonMain/kotlin/io/github/droidkaigi/confsched2019/data/api/response/ResponseImpl.kt
