+++ data/api-impl/src/main/java/io/github/droidkaigi/confsched2019/data/api/KtorDroidKaigiApi.kt
@@ -1,32 +1,10 @@
 package io.github.droidkaigi.confsched2019.data.api
 
-import io.github.droidkaigi.confsched2019.data.api.response.Response
-import io.github.droidkaigi.confsched2019.data.api.response.ResponseImpl
-import io.github.droidkaigi.confsched2019.data.api.response.SponsorResponse
-import io.github.droidkaigi.confsched2019.data.api.response.SponsorResponseImpl
 import io.ktor.client.HttpClient
-import io.ktor.client.request.accept
-import io.ktor.client.request.get
-import io.ktor.client.request.url
-import io.ktor.http.ContentType
 import javax.inject.Inject
 import javax.inject.Named
 
-class KtorDroidKaigiApi @Inject constructor(
-    val httpClient: HttpClient,
-    @Named("apiEndpoint") val apiEndpoint: String
-) : DroidKaigiApi {
-    override suspend fun getSponsors(): SponsorResponse {
-        return httpClient.get<SponsorResponseImpl> {
-            url("$apiEndpoint/sponsors")
-            accept(ContentType.Application.Json)
-        }
-    }
-
-    override suspend fun getSessions(): Response {
-        return httpClient.get<ResponseImpl> {
-            url("$apiEndpoint/timetable")
-            accept(ContentType.Application.Json)
-        }
-    }
-}
+class InjectableKtorDroidKaigiApi @Inject constructor(
+    httpClient: HttpClient,
+    @Named("apiEndpoint") apiEndpoint: String
+) : KtorDroidKaigiApi(httpClient, apiEndpoint)
